Handle hiding tokens with URLs
